Bugfixed bsp/Examples
git-svn-id: https://svn.illecker.at/hadoop/applications@25 f81d905f-b3a1-49e4-92e0-86bcef002278

+++ bsp/Examples/Examples.sh
@@ -1,9 +1,9 @@
 #!/bin/bash
 
-#if [ $# -ne 1 ]; then
-#  echo "Please specify main Class." 
-#  exit 1
-#fi
+if [ $# -lt 1 ]; then
+  echo "Please specify main Class." 
+  exit 1
+fi
 
 
 hadoop dfs -rmr output/examples
@@ -16,5 +16,5 @@ cd bin
 jar cf ../Examples.jar `find . -not -path "*/.svn/*" -not -type d` `find ../lib -not -path "*/.svn/*" -not -type d`
 cd ..
 
-hama jar Examples.jar $1
+hama jar Examples.jar $@
